Added pytest plugin for tensorized data and pytorch datasets
